+++ src/runtime_src/xdp/profile/device/hal_device/xdp_hal_device.h
@@ -42,7 +42,7 @@ public:
   virtual int getTraceBufferInfo(uint32_t nSamples, uint32_t& traceSamples, uint32_t& traceBufSz);
   virtual int readTraceData(void* traceBuf, uint32_t traceBufSz, uint32_t numSamples, uint64_t ipBaseAddress, uint32_t& wordsPerSample);
 
-  virtual uint32_t alloc(size_t sz, uint64_t memoryIndex);
+  virtual size_t alloc(size_t sz, uint64_t memoryIndex);
   virtual void free(uint32_t xdpBoHandle);
   virtual void* map(uint32_t xdpBoHandle);
   virtual void unmap(uint32_t xdpBoHandle);
